refactor character state into single state
Refactored code so that a single state handles all character objects,
and only one state needs to be updated instead of several.  Full game
functionality is completed, users can click on cards and a scoreboard
will keep track of the score, and the best score the user gets. CSS
styling is left for this project.